+++ log/2024-08-13-13-27-09.log
@@ -1,0 +1,8 @@
+
+[user] こんにちは
+
+[まさる](joy) やあ、みんな！今日はどんな楽しいことがあるかな？誰か面白い話とかない？俺、最近めっちゃ面白いこと見つけたんだけど、聞きたい？
+
+[きよこ](interest) あら、まさるくん元気ね！私も今日はテンション高いわ。何か楽しいことでもあったの？ぜひ聞かせてよ！
+
+[たかし](question) おう、まさる。何か面白いこと見つけたのか？俺も気になるな。ただ、あんまり長くならないようにしてくれよ。あとで用事があるんだ。
