+++ log/2024-08-13-13-51-07.log
@@ -1,0 +1,19 @@
+
+[user] 歌いましょう
+
+[まさる](joy)よし、みんなで歌おう！カラオケに行くのはどうだ？盛り上がること間違いなしだぞ。誰か得意な曲とかある？
+
+[きよこ](interest)カラオケいいね！私、最近のJ-POPが好きなんだ。みんなで盛り上がれる曲を歌いたいな。まさるは何か歌いたい曲あるの？
+
+[たかし](surprise)え、急にカラオケ？まあ、いいけど...。僕はあまり上手じゃないんだよね。でも、みんなと一緒なら楽しめそう。誰か昭和の名曲とか知ってる？
+
+[まさる](joy)おー、たかし！昭和の曲いいね。僕も好きだぞ。「上を向いて歩こう」なんてどうだ？みんなで歌えば盛り上がるぞ！
+
+[きよこ](interest)えー、昭和の曲？ちょっと古いかな。でも、みんなで歌えば楽しそう！じゃあ、新しい曲と昭和の曲をミックスしてみるのはどう？
+[user] 　　　　quit
+
+[まさる](surprise)え？もう終わりなのか？せっかく盛り上がってきたところなのに...。でも、みんなの予定もあるだろうし、仕方ないか。また近いうちにカラオケ行こうぜ！
+
+[きよこ](question)quit？もう終わっちゃうの？楽しくなってきたところだったのに残念。でも、次はもっと長く歌える日を作ろうよ。みんなのスケジュール合わせて計画しよう！
+
+[たかし](interest)そうか、もう終わりか。実は少し緊張してたんだ。でも、みんなと歌うのは意外と楽しかったな。次はもっとリラックスして参加できそうだ。また誘ってくれるといいな。
